+++ assets/js/modules/private_channel.js
@@ -9,9 +9,6 @@ export default ( dispatch ) => {
   // ignore privateChannel when no valid token and return unjoined
   if(window.userToken === '') return channel
 
-  // listen for events
-
-
   // join channel
   channel.join()
   .receive('ok', res => {