+++ assets/js/modules/session/index.js
@@ -1,6 +1,5 @@
 import reducer from "./reducers";
 
 export { default as sessionSelectors } from "./selectors";
-export { default as sessionOperations } from "./operations";
 
 export default reducer;