+++ assets/js/modules/session/types.js
@@ -1,4 +1,3 @@
-const CHANGE_USERNAME = "sesson/CHANGE_USERNAME"
 const LOGIN = 'session/LOGIN'
 
-export default { CHANGE_USERNAME, LOGIN }
+export default { LOGIN }