+++ assets/js/pages/index.js
@@ -1,19 +1,11 @@
-import About from './about'
 import Error from './error'
-import Login from './login'
 import Scene from './scene'
 import Simulation from './simulation.js'
-import Settings from './settings'
 import VideoAdd from './video.add'
-import VideoList from './video.list'
 
 export default {
-  About,
   Error,
-  Login,
   Scene,
   Simulation,
-  Settings,
-  VideoAdd,
-  VideoList
+  VideoAdd
 }