+++ assets/js/views/loader.js
@@ -4,9 +4,9 @@ export default viewPath => {
   let view
   try {
     const ViewClass = require('./' + viewPath).default
-    view = new ViewClass()
+    view = new ViewClass.default
   } catch (e) {
-    view = new MainView()
+    view = new MainView
   }
   return view
 }