+++ lib/crowd_crush/accounts/user.ex
@@ -4,14 +4,9 @@ defmodule CrowdCrush.Accounts.User do
 
   schema "users" do
     field :username, :string
-    has_one :credential, Credential
+    field :password, :string, virtual: true
+    field :password_hash, :string
     timestamps()
-  end
-
-  def registration_changeset(user, params) do
-    user
-    |> changeset(params)
-    |> cast_assoc(:credential, with: &Credential.changeset/2, required: true)
   end
 
   def changeset(user, attrs) do
@@ -19,5 +14,24 @@ defmodule CrowdCrush.Accounts.User do
     |> cast(attrs, [:username])
     |> validate_required([:username])
     |> validate_length(:username, min: 3, max: 20)
+  end
+
+  def registration_changeset(user, params) do
+    user
+    |> changeset(params)
+    |> cast(params, [:password])
+    |> validate_required([:password])
+    |> validate_length(:password, min: 6, max: 100)
+    |> put_pass_hash()
+  end
+
+  defp put_pass_hash(changeset) do
+    case changeset do
+      %Ecto.Changeset{valid?: true, changes: %{password: pass}} ->
+        put_change(changeset, :password_hash, Pbkdf2.hash_pwd_salt(pass))
+
+      _ ->
+        changeset
+    end
   end
 end