+++ lib/crowd_crush_web/controllers/page_controller.ex
@@ -1,7 +1,13 @@
 defmodule CrowdCrushWeb.PageController do
   use CrowdCrushWeb, :controller
 
-  def index(conn, _), do: render conn, "index.html"
+  def about(conn, _), do: render conn, "about.html"
+
+  def index(conn, _) do
+    if is_nil(conn.assigns.current_user),
+      do: redirect(conn, to: Routes.page_path(conn, :about)),
+      else: redirect(conn, to: Routes.video_path(conn, :index))
+  end
 
   def test(conn, _), do: render conn, "test.html"
 end