+++ lib/crowd_crush_web/views/session_view.ex
@@ -1,0 +1,3 @@
+defmodule CrowdCrushWeb.SessionView do
+  use CrowdCrushWeb, :view
+end