+++ lib/crowd_crush_web/views/view_helpers.ex
@@ -1,0 +1,10 @@
+defmodule CrowdCrushWeb.ViewHelpers do
+  @moduledoc """
+  Conveniences for all views.
+  """
+  use Phoenix.HTML
+
+  def icon(name, opts \\ [] ), do: content_tag(:i, "",
+    [{:class, "icon-#{name} #{Keyword.get(opts, :class, "")}"} | Keyword.delete(opts, :class)])
+
+end